CompanyListFragment をEpoxy表示のリファクタリング
