+++ app/src/main/res/layout/fragment_liner_list_view.xml
@@ -7,11 +7,6 @@
     <data>
 
         <variable
-            name="portName"
-            type="String"
-            />
-
-        <variable
             name="showableComment"
             type="Boolean"
             />
@@ -21,77 +16,92 @@
             type="com.ikmr.banbara23.yaeyama_liner_checker.model.PortStatus"
             />
 
+        <variable
+            name="onItemViewClick"
+            type="View.OnClickListener"
+            />
+
         <import type="com.ikmr.banbara23.yaeyama_liner_checker.common.StatusHelper" />
 
         <import type="android.view.View" />
     </data>
 
-    <RelativeLayout
+    <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/white"
-        android:paddingLeft="16dp"
-        android:paddingTop="2dp"
-        android:paddingRight="16dp"
-        android:paddingBottom="2dp"
+        android:background="?android:attr/selectableItemBackground"
+        android:clickable="true"
+        android:focusable="true"
+        android:onClick="@{onItemViewClick}"
         >
 
-        <LinearLayout
+        <RelativeLayout
             android:layout_width="match_parent"
-            android:layout_height="56dp"
-            android:layout_toStartOf="@+id/view_status_list_status"
-            android:layout_toLeftOf="@+id/view_status_list_status"
-            android:gravity="center_vertical"
-            android:orientation="vertical"
+            android:layout_height="wrap_content"
+            android:background="@color/white"
+            android:paddingLeft="16dp"
+            android:paddingTop="2dp"
+            android:paddingRight="16dp"
+            android:paddingBottom="2dp"
             >
 
-            <!--港-->
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="56dp"
+                android:layout_toStartOf="@+id/view_status_list_status"
+                android:layout_toLeftOf="@+id/view_status_list_status"
+                android:gravity="center_vertical"
+                android:orientation="vertical"
+                >
+
+                <!--港-->
+                <TextView
+                    android:id="@+id/view_status_list_port_text"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:gravity="center_vertical"
+                    android:text="@{portStatus.portName}"
+                    android:textColor="@color/secondary_text"
+                    android:textSize="18dp"
+                    tools:text="竹富"
+                    />
+
+                <!--コメント-->
+                <TextView
+                    android:id="@+id/view_status_list_description_text"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="8dp"
+                    android:ellipsize="end"
+                    android:gravity="center_vertical"
+                    android:maxLines="1"
+                    android:text="@{portStatus.comment}"
+                    android:textColor="@color/dark_grey"
+                    android:textSize="14dp"
+                    android:visibility="@{showableComment ? View.VISIBLE : View.GONE}"
+                    tools:text="コメントコメントコメントコメントコメント"
+                    tools:visibility="visible"
+                    />
+
+            </LinearLayout>
+            <!--ステータス文字-->
             <TextView
-                android:id="@+id/view_status_list_port_text"
+                android:id="@+id/view_status_list_status"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:gravity="center_vertical"
-                android:text="@{portStatus.portName}"
-                android:textColor="@color/secondary_text"
-                android:textSize="18dp"
-                tools:text="竹富"
+                android:layout_alignParentEnd="true"
+                android:layout_alignParentRight="true"
+                android:layout_centerVertical="true"
+                android:background="@{StatusHelper.getStatusColor(portStatus.status.code)}"
+                android:paddingLeft="4dp"
+                android:paddingTop="2dp"
+                android:paddingRight="4dp"
+                android:paddingBottom="2dp"
+                android:text="@{portStatus.status.text}"
+                android:textColor="@color/white"
+                android:textSize="16dp"
+                tools:text="一部欠航便有り"
                 />
-
-            <!--コメント-->
-            <TextView
-                android:id="@+id/view_status_list_description_text"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:ellipsize="end"
-                android:gravity="center_vertical"
-                android:maxLines="1"
-                android:text="@{portStatus.comment}"
-                android:textColor="@color/dark_grey"
-                android:textSize="14dp"
-                android:visibility="@{showableComment ? View.VISIBLE : View.GONE}"
-                tools:text="コメントコメントコメントコメントコメント"
-                tools:visibility="visible"
-                />
-
-        </LinearLayout>
-        <!--ステータス文字-->
-        <TextView
-            android:id="@+id/view_status_list_status"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:layout_alignParentRight="true"
-            android:layout_centerVertical="true"
-            android:background="@{StatusHelper.getStatusColor(portStatus.status.code)}"
-            android:paddingLeft="4dp"
-            android:paddingTop="2dp"
-            android:paddingRight="4dp"
-            android:paddingBottom="2dp"
-            android:text="@{portStatus.status.text}"
-            android:textColor="@color/white"
-            android:textSize="16dp"
-            tools:text="一部欠航便有り"
-            />
-    </RelativeLayout>
+        </RelativeLayout>
+    </FrameLayout>
 </layout>
